requirements.txt
langchain==0.1.16
langchain-core==0.1.48
langchain-experimental==0.0.57
openai==1.3.0
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.4
seaborn==0.13.2
python-dotenv==1.0.0
pydantic==2.5.0

